PLANNER-2116 Externalize build method of ForagerConfig
